+++ android/src/main/kotlin/uz/shs/better_player_plus/BetterPlayer.kt
@@ -47,6 +47,7 @@ import androidx.media3.common.Timeline
 import androidx.media3.common.Tracks
 import androidx.media3.common.TrackSelectionOverride
 import androidx.media3.common.VideoSize
+import androidx.media3.common.text.CueGroup
 import androidx.media3.common.util.UnstableApi
 import androidx.media3.common.util.Util
 import androidx.media3.datasource.DataSource
@@ -545,6 +546,14 @@ internal class BetterPlayer(
 
             override fun onVideoSizeChanged(videoSize: VideoSize) {
                 sendVideoSizeChanged(videoSize)
+            }
+
+            override fun onCues(cueGroup: CueGroup) {
+                val event: MutableMap<String, Any?> = HashMap()
+                event["event"] = "cuesChanged"
+                event["key"] = key
+                event["cues"] = cueGroup.cues.mapNotNull { it.text?.toString() }
+                eventSink.success(event)
             }
 
             override fun onTracksChanged(tracks: Tracks) {